HQNetwork

+++ utils/replay_memory.py
@@ -12,7 +12,7 @@ class ReplayMemory(object):
 		#use memory maps so we won't have to worry about eating up lots of RAM
 		get_path = lambda name: os.path.join(dirname, name)
 		self.screens = np.memmap(get_path('screens'), dtype=np.float32, mode='w+', shape=tuple([self.maxlen]+input_shape))
-		self.phases = np.memmap(get_path('phases'), dtype=np.int16, mode='w+', shape=(self.maxlen,))
+		self.phases = np.memmap(get_path('phases'), dtype=np.int32, mode='w+', shape=(self.maxlen,))
 		self.actions = np.memmap(get_path('actions'), dtype=np.float32, mode='w+', shape=(self.maxlen, action_size))
 		self.rewards = np.memmap(get_path('rewards'), dtype=np.float32, mode='w+', shape=(self.maxlen,))
 		self.is_terminal = np.memmap(get_path('terminals'), dtype=np.bool, mode='w+', shape=(self.maxlen,))
